ml_backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
scikit-learn
pyportfolioopt
python-dotenv
google-generativeai
python-multipart